Fix note off problem. Closes #4
This was caused because I was filtering NPRNs out in a stupid way.
Changed to using midiselect to ignore them and everything works properly
again.

Background: Live records the cc's associated with nprn's into clips.
There's no way to avoid this so we need to filter them out as we're
handling them directly with the lh_midi stuff and things would get very
confusing if we sent them back out to the mopho.